refactor: position 패키지 생성
