+++ src/main/java/janggi/domain/piece/Guard.java
@@ -1,5 +1,7 @@
 package janggi.domain.piece;
 
+import janggi.domain.position.Position;
+
 import java.util.List;
 
 public class Guard extends Piece {
