+++ src/main/java/janggi/domain/piece/King.java
@@ -1,5 +1,7 @@
 package janggi.domain.piece;
 
+import janggi.domain.position.Position;
+
 import java.util.List;
 
 public class King extends Piece {
