+++ src/main/java/janggi/domain/piece/Pieces.java
@@ -1,5 +1,7 @@
 package janggi.domain.piece;
 
+import janggi.domain.position.Position;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
