+++ src/test/java/janggi/domain/PositionTest.java
@@ -1,6 +1,6 @@
 package janggi.domain;
 
-import janggi.domain.piece.Position;
+import janggi.domain.position.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
