+++ src/test/java/janggi/domain/piece/PieceTest.java
@@ -1,6 +1,7 @@
 package janggi.domain.piece;
 
 import janggi.domain.ReplaceUnderBar;
+import janggi.domain.position.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
