+++ src/test/java/janggi/domain/piece/PiecesTest.java
@@ -1,6 +1,7 @@
 package janggi.domain.piece;
 
 import janggi.domain.ReplaceUnderBar;
+import janggi.domain.position.Position;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
