+++ src/test/java/janggi/domain/piece/RookTest.java
@@ -1,5 +1,6 @@
 package janggi.domain.piece;
 
+import janggi.domain.position.Position;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
